Clean and Fix GCC9 regression
Clean the code style and fix a regression with GCC9: it seems that GCC9
generates and ICE if the object xml_tag has a defined filazer. Now there
is a preprocessing flag that check the version of GCC and if it is newer
than 8.x trunk the finalizer is trimmed out.